Create custom RSpec matcher to validate form group HTML output
The specs that validate what kind of content is generated for all
the supported HTML tag built through the Form Builder have a lot
of repeated code.

This commit tackles that issue by supporting a custom RSpec matcher
that validates the HTML output generated by the Form Builder.

This commit does NOT change any behaviour in the way that the Form
Build works.